scitbx/lbfgs/raw.h: changes by Randy Read:   1. if diagco != 0: stp = 1   2. if diagco == 3: "Second" below
There are two changes.  First, if you were using the proper inverse  
Hessian in Newton's method, the first trial in a line search would  
have a step length of one, with the search direction being given by  
gradient times inverse Hessian.  Similarly, if you're using  
approximate Hessian information in (L-)BFGS, the expected initial step  
length should be one.  If there's no Hessian information, then it  
makes sense to have a rule of thumb that gives a sensible initial step  
(i.e. 1/sqrt(norm(gradient))).  So the first change is to make stp1  
equal to 1 when Hessian information is being used (diagco!=0).  With  
this change, the initial step is nearly as good as the one from the  
Newton's method optimization.

Second, the ys/yy scaling of the gradient only makes sense, I think,  
if the gradient isn't being scaled by the inverse curvatures before  
being used as a search direction. I've changed it so the y terms are  
multiplied by the corresponding diag elements.  I'm not sure this is  
correct -- it works better than the original ys/yy scaling, but not  
nearly as well as applying no scaling to the initial inverse Hessian.   
Also, it leads to an incorrect search direction, so there must be  
something wrong with the reasoning.