Change design of restraint parser to fit parser/builder pattern.
In order to pass information from atom_parser to restraint_parser,
a handful of structures were declared as class attributes of the
former while the latter inherited from the former. However, in
a parser/builder pattern, the builder shall be use to share such
information. Moreover this designed was not thread-safe (two parser
instances each in its own thread would share those class attributes).

Thus, enrich the API of the builder, and change the two parsers
to take advantage of it.