First pass at using XPP gain maps.
Metrology isn't used correctly yet.  More to come.

1) Modifies convert_gain_map to read an XPP ASCII gain map and write
it to a pickle file.  This file is displayable by cctbx.image_viewer.

2) Adds gain_map_level to the pyana module common_mode.  If present,
instead of using the supplied gain map directly, the gain level is
applied instead in a binary fashion, based on whether the pixels in
the gain map are zeros or ones.