cxi.mpi_submit knows about the LCLS queuing system, SGE queuing systems and the NERSC PBS queuing system. However, some systems are too complex for this program and need a lot of special environment variables for jobs to run on their cluster. To this end, add the mp option 'custom' for processing xfel data with a specific subumission script.
cxi.mpi_submit will look for <command>, <queue> and <nprocs> in the
script and replace them as appropiate. <command> is replaced with the
full command including the dispatcher requested for processing the xtc
streams.

Also add a dry_run parameter. If True, the program will create the trial
directory prepare the config, phil, and submission files for the run,
but not submit the job, instead printing the command that would submit
the job to stdout.