Two bugfixes for cxi.merge sdfac refinement in apply_sd_error_params()
First, in the step where the corrected sigma is calculated, avoid square roots
of negatives by throttling the value in the square root to a minimum, according
to Evans 2013.

Second, when computing I/sigma, return the original I over sigma instead of
the scaled I/sigma, which is what the calling program expects.