Fix libtbx.import_all_ext.  This test imports all .so files it finds in the build folder one at a time to test that they can be imported.  This fails when a c++ extension has dependencies on another.  For this reason, we don't generally import modules this way when we actually run our software.
To preserve the functionality of the test, add a step that walks the tree of installed modules (libtbx, scitbx, cctbx, etc.), looking for __init__.py files.  When it finds them, it imports the associated module (eg import scitbx.array_family).  This allows the code in the __init__.py files to be ran, which sets up the boost extentions properly.

Then the .so files can be loaded as part of the test.

As part of this process, several missing __init__.py files have been added.