Propagate stack traces through easy_mp.
Currently, if an easy_mp task fails, the stack trace points at the line in
scheduling/result.py, where the exception is re-raised. This is obviously not
helpful for debugging. Instead try to extend the Exception object and store a
copy of the stack trace alongside, which can be unpacked to re-raise later.
As the Exception object may undergo pickling/unpickling, and stack traces are
un-pickleable, keep a second, preformatted copy of the stack trace around
which can be presented to the user/developer instead.

Also add a test to see if the original cause of any Exception is reported.